% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isorropia_variables_helper.R
\name{isorropia_variables_helper}
\alias{isorropia_variables_helper}
\title{Function to load a table with information about the variables used in the 
ISORROPIA II model.}
\usage{
isorropia_variables_helper()
}
\value{
Tibble.
}
\description{
\code{isorropia_variables_helper} is helpful for joining and decoding the 
model's variables.
}
\author{
Stuart K. Grange
}
