% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isorropia_demo_data.R
\name{read_isorropia_demo_data}
\alias{read_isorropia_demo_data}
\title{Function to read/load the demo data that is supplied with the ISORROPIA II
code package.}
\usage{
read_isorropia_demo_data()
}
\value{
Tibble.
}
\description{
Function to read/load the demo data that is supplied with the ISORROPIA II
code package.
}
\author{
ISORROPIA's developers.
}
