% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isorropia_output_file.R
\name{read_isorropia_output_file}
\alias{read_isorropia_output_file}
\alias{read_isorropia_input_file}
\title{Functions to read input and output data files used by the ISORROPIA II model.}
\usage{
read_isorropia_output_file(file)

read_isorropia_input_file(file, unnest = FALSE)
}
\arguments{
\item{file}{File name of an ISORROPIA II input or output data file.}

\item{unnest}{For \code{read_isorropia_input_file}, should the return be 
unnested?}
}
\value{
Tibble.
}
\description{
Functions to read input and output data files used by the ISORROPIA II model.
}
\author{
Stuart K. Grange
}
