% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_isorropia.R
\name{run_isorropia}
\alias{run_isorropia}
\title{Function to run the ISORROPIA II model and return the model's input and 
output in a friendly format.}
\usage{
run_isorropia(
  df,
  directory_isorropia,
  problem_type = "forward",
  verbose = FALSE
)
}
\arguments{
\item{df}{Input data frame/tibble that is to be modelled with ISORROPIA II.}

\item{directory_isorropia}{Directory where ISORROPIA II's programme is 
located. The programme can either be the complied source (\code{isorropia}) 
or the provided Windows \code{.exe} file (\code{isrpia2.exe}).}

\item{problem_type}{Problem type for ISORROPIA II to solve. 
\code{problem_type} can either be \code{"forward"} or \code{"reverse"}.}

\item{verbose}{Should the function give messages?}
}
\value{
Nested tibble with a single row.
}
\description{
Function to run the ISORROPIA II model and return the model's input and 
output in a friendly format.
}
\examples{

\dontrun{

# Load demo data
data_demo <- read_isorropia_demo_data()

# Run the ISORROPIA II model with demo data using a complied version of 
# isorropia
run_isorropia(
  data_demo,
  directory_isorropia = "~/isorropia/source",
  verbose = TRUE
)

# Run ISORROPIA II with an executable file on a Windows system
run_isorropia(
  data_demo,
  directory_isorropia = "C:/isorropia/ISO2_1Bin",
  verbose = TRUE
)

}

}
\author{
Stuart K. Grange
}
