% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_isorropia_variables.R
\docType{data}
\name{select_isorropia_variables}
\alias{select_isorropia_variables}
\alias{variables_isorropia}
\title{Function to select the variables that are required for the ISORROPIA II 
model.}
\format{
An object of class \code{character} of length 10.
}
\usage{
select_isorropia_variables(df, drop = FALSE)

variables_isorropia
}
\arguments{
\item{df}{Data frame or tibble containing the required variables.}

\item{drop}{Should the ISORROPIA II variables be dropped rather than kept?}
}
\value{
Data frame or tibble
}
\description{
Function to select the variables that are required for the ISORROPIA II 
model.
}
\seealso{
\code{\link{read_isorropia_demo_data}}
}
\author{
Stuart K. Grange
}
\keyword{datasets}
