% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_re_exports.R
\name{utils-functions}
\alias{utils-functions}
\title{Pseudo-function to re-export functions from the \strong{utils} package.}
\arguments{
\item{utils}{functions}
}
\description{
Pseudo-function to re-export functions from the \strong{utils} package.
}
