% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_isorropia.R
\name{run_isorropia}
\alias{run_isorropia}
\title{Function to run the ISORROPIA II model and return the model's input and 
output in a friendly format.}
\usage{
run_isorropia(df, directory_isorropia, verbose = FALSE)
}
\arguments{
\item{df}{Input data frame/tibble that is to be modelled with ISORROPIA II.}

\item{directory_isorropia}{Directory where ISORROPIA II's programme is 
located.}

\item{verbose}{Should the function give messages?}
}
\value{
Nested tibble with a single row.
}
\description{
Function to run the ISORROPIA II model and return the model's input and 
output in a friendly format.
}
\examples{

\dontrun{

# Load demo data
data_demo <- read_isorropia_demo_data()

# Run the ISORROPIA II model with demo data
run_isorropia(
  data_demo,
  directory_isorropia = "~/isorropia/source",
  verbose = TRUE
)

}

}
\author{
Stuart K. Grange
}
